feat: add options.skipEncoding() to pug-load.read()
This allows Pug to correctly read non-utf8 files, if the user wants to
do such a thing.

The optional function takes the same arguments as load.read(), and
should return a truthy/falsey value. If truthy, load.read() will NOT
specify an encoding when it reads the file from disk. Otherwise, it will
continue to specify "utf-8" as the encoding.